requirements.txt
pandas>=2.0.0
flask>=3.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
